Finetuning

+++ post_hoc_cbm/training_tools/embedding_tools.py
@@ -36,6 +36,28 @@ def get_projections(args, backbone, posthoc_layer, loader):
             all_lbls = np.concatenate([all_lbls, batch_Y.numpy()], axis=0)
     return all_embs, all_projs, all_lbls
 
+def get_projections_ft(args, backbone, posthoc_layer, loader):
+    all_projs, all_embs, all_lbls = None, None, None
+    for batch in tqdm(loader):
+        batch_X, batch_Y = unpack_batch(batch)
+        batch_X = batch_X.to(args.device)
+        #batch_X = preprocess(batch_X) ############################################# new because we load the images without preprocessing
+        if "clip" in args.backbone_name:
+            embeddings = backbone.encode_image(batch_X).detach().float()
+        else:
+            embeddings = backbone(batch_X).detach()
+        projs = posthoc_layer.compute_dist(embeddings).detach().cpu().numpy()
+        embeddings = embeddings.detach().cpu().numpy()
+        if all_embs is None:
+            all_embs = embeddings
+            all_projs = projs
+            all_lbls = batch_Y.numpy()
+        else:
+            all_embs = np.concatenate([all_embs, embeddings], axis=0)
+            all_projs = np.concatenate([all_projs, projs], axis=0)
+            all_lbls = np.concatenate([all_lbls, batch_Y.numpy()], axis=0)
+    return all_embs, all_projs, all_lbls
+
 
 class EmbDataset(Dataset):
     def __init__(self, data, target):
@@ -81,6 +103,53 @@ def load_or_compute_projections(args, backbone, posthoc_layer, train_loader, tes
     else:
         train_embs, train_projs, train_lbls = get_projections(args, backbone, posthoc_layer, train_loader)
         test_embs, test_projs, test_lbls = get_projections(args, backbone, posthoc_layer, test_loader)
+
+        np.save(train_file, train_embs)
+        np.save(test_file, test_embs)
+        np.save(train_proj_file, train_projs)
+        np.save(test_proj_file, test_projs)
+        np.save(train_lbls_file, train_lbls)
+        np.save(test_lbls_file, test_lbls)
+    
+    return train_embs, train_projs, train_lbls, test_embs, test_projs, test_lbls
+
+
+def load_or_compute_projections_ft(args, backbone, posthoc_layer, train_loader, test_loader):
+    # Get a clean conceptbank string
+    # e.g. if the path is /../../cub_resnet-cub_0.1_100.pkl, then the conceptbank string is resnet-cub_0.1_100
+    conceptbank_source = args.concept_bank.split("/")[-1].split(".")[0] 
+    
+    # To make it easier to analyize results/rerun with different params, we'll extract the embeddings and save them
+
+    ################################# changed the name to _ft to make sure that the projections are computed again for the finetuned clip model 
+    train_file = f"train-embs_ft{args.dataset}__{args.backbone_name}__{conceptbank_source}.npy"
+    test_file = f"test-embs_ft{args.dataset}__{args.backbone_name}__{conceptbank_source}.npy"
+    train_proj_file = f"train-proj_ft{args.dataset}__{args.backbone_name}__{conceptbank_source}.npy"
+    test_proj_file = f"test-proj_ft{args.dataset}__{args.backbone_name}__{conceptbank_source}.npy"
+    train_lbls_file = f"train-lbls_ft{args.dataset}__{args.backbone_name}__{conceptbank_source}_lbls.npy"
+    test_lbls_file = f"test-lbls_ft{args.dataset}__{args.backbone_name}__{conceptbank_source}_lbls.npy"
+    
+
+    train_file = os.path.join(args.out_dir, train_file)
+    test_file = os.path.join(args.out_dir, test_file)
+    train_proj_file = os.path.join(args.out_dir, train_proj_file)
+    test_proj_file = os.path.join(args.out_dir, test_proj_file)
+    train_lbls_file = os.path.join(args.out_dir, train_lbls_file)
+    test_lbls_file = os.path.join(args.out_dir, test_lbls_file)
+
+    if os.path.exists(train_proj_file):
+        print("Projections exist.")
+        train_embs = np.load(train_file)
+        test_embs = np.load(test_file)
+        train_projs = np.load(train_proj_file)
+        test_projs = np.load(test_proj_file)
+        train_lbls = np.load(train_lbls_file)
+        test_lbls = np.load(test_lbls_file)
+
+    else:
+        print("Projections do not exist.")
+        train_embs, train_projs, train_lbls = get_projections_ft(args, backbone, posthoc_layer, train_loader)
+        test_embs, test_projs, test_lbls = get_projections_ft(args, backbone, posthoc_layer, test_loader)
 
         np.save(train_file, train_embs)
         np.save(test_file, test_embs)
